database-tests/app/requirements.txt
redis==3.5.3
pymongo==3.12.0
cassandra-driver==3.25.0
scylla-driver==3.26.3